+++ playground/src/playground.h
@@ -8,5 +8,5 @@
 
 b8 playground_on_initialize(struct program* program);
 b8 playground_on_deinitialize(struct program* program);
-b8 playground_on_tick(struct program* program, f64 delta_time);
-b8 playground_on_frame(struct program* program, f64 delta_time);
+b8 playground_on_tick(struct program* program, u64 delta_time);
+b8 playground_on_frame(struct program* program, u64 delta_time);
